+++ test/testflows/helpers/cluster.py
@@ -1,15 +1,14 @@
-import os
-import time
-import inspect
-import threading
-import tempfile
-
 import glob
-import testflows.settings as settings
+import inspect
+import os
+import tempfile
+import threading
+import time
 
-from testflows.core import *
+import testflows.settings as settings
 from testflows.asserts import error
 from testflows.connect import Shell as ShellBase
+from testflows.core import *
 from testflows.uexpect import ExpectTimeoutError
 
 MESSAGES_TO_RETRY = [
@@ -689,7 +688,7 @@ class Cluster(object):
                     with And("executing docker-compose up"):
                         for up_attempt in range(max_up_attempts):
                             with By(f"attempt {up_attempt}/{max_up_attempts}"):
-                                cmd = self.command(None, f"{self.docker_compose} up --renew-anon-volumes --force-recreate --timeout 120 -d 2>&1 | tee", timeout=timeout)
+                                cmd = self.command(None, f"{self.docker_compose} up --timeout 120 -d 2>&1 | tee", timeout=timeout)
                                 if "is unhealthy" not in cmd.output:
                                     break